3 Top children’s book illustrators
3 Top children’s book illustrators 

E.H Shepard, Leo Lionni and Ezra Jack are one of the <a href="https://childrensbookillustrators.net/">top children’s book illustrators</a>. Their work is a must follow, the ultimate guru’s for someone who is new to this field. 

E.H Shepard, who is the creator of the famous honey loving characters named Pooh from Winnie-The-Pooh. E.H worked on several famous stories such as David Copperfield, Aesop’s Fables and Tom Browns Schooldays. From there onwards he was well recognized within the industry, which lead to his prominent success as an illustrator. 

Loe Lionni, finessed in the field of art as a painter and illustrator; during which he produced more than 40 books for children. He received various awards for his work, one of which was an; American Institute of Graphic Arts (A.I.G.A.) Gold Medal.

Ezra Jack, who won the 1963 Caldecott Medal for illustrating The Snowy Day, which he himself has written; he was a multitalented illustrator just like the rest in the list. His illustration turned his life around, whilst working on an illustration and story where a black kid was the protagonist. He illustrated a total of 22 books, out of which turned out to children’s classics. 

Which of these illustrators influenced you the most?